+++ optbinning/__init__.py
@@ -4,8 +4,11 @@ from .binning.multiclass_binning import MulticlassOptimalBinning
 
 from .binning.binning_process import BinningProcess
 
+from .binning.mdlp import MDLP
+
 
 __all__ = ['OptimalBinning',
            'ContinuousOptimalBinning',
            'MulticlassOptimalBinning',
-           'BinningProcess']
+           'BinningProcess',
+           'MDLP']