+++ optbinning/binning/multiclass_cp.py
@@ -35,7 +35,7 @@ class MulticlassBinningCP(BinningCP):
         self._n = None
         self._x = None
 
-    def build_model(self, n_nonevent, n_event):
+    def build_model(self, n_nonevent, n_event, trend_changes):
         # Parameters
         M = int(1e6)
         D, V, pvalue_violation_indices = multiclass_model_data(
@@ -90,6 +90,14 @@ class MulticlassBinningCP(BinningCP):
                 else:
                     self.add_constraint_monotonic_valley(
                         model, n, D[c], x, c, y, M)
+
+            elif self.monotonic_trend == "peak_heuristic":
+                self.add_constraint_monotonic_peak_heuristic(
+                    model, n, D[c], x, trend_changes[c], M)
+
+            elif self.monotonic_trend == "valley_heuristic":
+                self.add_constraint_monotonic_valley_heuristic(
+                    model, n, D[c], x, trend_changes[c], M)
 
         # constraint: max-pvalue
         for c in range(n_classes):