+++ tests/test_binning.py
@@ -5,6 +5,7 @@ OptimalBinning testing.
 # Guillermo Navas-Palencia <g.navas.palencia@gmail.com>
 # Copyright (C) 2020
 
+import numpy as np
 import pandas as pd
 
 from pytest import approx, raises
@@ -161,19 +162,149 @@ def test_numerical_default():
                                   13.70499992, 15.04500008, 16.92500019],
                                  rel=1e-6)
 
+    optb.binning_table.build()
+    assert optb.binning_table.iv == approx(5.04392547, rel=1e-6)
+
+    optb.binning_table.analysis()
+    assert optb.binning_table.gini == approx(0.87541620, rel=1e-6)
+    assert optb.binning_table.js == approx(0.39378376, rel=1e-6)
+    assert optb.binning_table.quality_score == approx(0.0, rel=1e-6)
+
+    with raises(ValueError):
+        optb.binning_table.plot(metric="new_metric")
+
 
 def test_numerical_default_solvers():
     optb_mip_cbc = OptimalBinning(solver="mip", mip_solver="cbc")
-    optb_mip_cbc.fit(x, y)
-
     optb_mip_bop = OptimalBinning(solver="mip", mip_solver="bop")
-    optb_mip_bop.fit(x, y)
-
     optb_cp = OptimalBinning(solver="cp")
-    optb_cp.fit(x, y)
 
     for optb in [optb_mip_bop, optb_mip_cbc, optb_cp]:
+        optb.fit(x, y)
         assert optb.status == "OPTIMAL"
+        assert optb.splits == approx([11.42500019, 12.32999992, 13.09499979,
+                                      13.70499992, 15.04500008, 16.92500019],
+                                     rel=1e-6)
+
+
+def test_numerical_user_splits():
+    user_splits = [11, 12, 13, 14, 15, 17]
+    optb = OptimalBinning(user_splits=user_splits, max_pvalue=0.05)
+    optb.fit(x, y)
+
+    assert optb.status == "OPTIMAL"
+    assert optb.splits == approx([13, 15, 17], rel=1e-6)
+
+    optb.binning_table.build()
+    assert optb.binning_table.iv == 4.819661314733627
+
+    optb = OptimalBinning(user_splits=user_splits, max_pvalue=0.05,
+                          max_pvalue_policy="all")
+    optb.fit(x, y)
+    optb.binning_table.build()
+    assert optb.binning_table.iv == 4.819661314733627
+
+
+def test_categorical_default_user_splits():
+    x = np.array([
+        'Working', 'State servant', 'Working', 'Working', 'Working',
+        'State servant', 'Commercial associate', 'State servant',
+        'Pensioner', 'Working', 'Working', 'Pensioner', 'Working',
+        'Working', 'Working', 'Working', 'Working', 'Working', 'Working',
+        'State servant', 'Working', 'Commercial associate', 'Working',
+        'Pensioner', 'Working', 'Working', 'Working', 'Working',
+        'State servant', 'Working', 'Commercial associate', 'Working',
+        'Working', 'Commercial associate', 'State servant', 'Working',
+        'Commercial associate', 'Working', 'Pensioner', 'Working',
+        'Commercial associate', 'Working', 'Working', 'Pensioner',
+        'Working', 'Working', 'Pensioner', 'Working', 'State servant',
+        'Working', 'State servant', 'Commercial associate', 'Working',
+        'Commercial associate', 'Pensioner', 'Working', 'Pensioner',
+        'Working', 'Working', 'Working', 'Commercial associate', 'Working',
+        'Pensioner', 'Working', 'Commercial associate',
+        'Commercial associate', 'State servant', 'Working',
+        'Commercial associate', 'Commercial associate',
+        'Commercial associate', 'Working', 'Working', 'Working',
+        'Commercial associate', 'Working', 'Commercial associate',
+        'Working', 'Working', 'Pensioner', 'Working', 'Pensioner',
+        'Working', 'Working', 'Pensioner', 'Working', 'State servant',
+        'Working', 'Working', 'Working', 'Working', 'Working',
+        'Commercial associate', 'Commercial associate',
+        'Commercial associate', 'Working', 'Commercial associate',
+        'Working', 'Working', 'Pensioner'], dtype=object)
+
+    y = np.array([
+        1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0])
+
+    optb = OptimalBinning(dtype="categorical", solver="mip", cat_cutoff=0.1,
+                          verbose=True)
+    optb.fit(x, y)
+
+    assert optb.status == "OPTIMAL"
+
+    user_splits = [['Pensioner', 'Working'],
+                   ['Commercial associate'], ['State servant']]
+
+    optb = OptimalBinning(dtype="categorical", solver="mip", cat_cutoff=0.1,
+                          user_splits=user_splits, verbose=True)
+    optb.fit(x, y)
+
+    assert optb.status == "OPTIMAL"
+
+
+def test_auto_modes():
+    optb0 = OptimalBinning(monotonic_trend="auto")
+    optb1 = OptimalBinning(monotonic_trend="auto_heuristic")
+    optb2 = OptimalBinning(monotonic_trend="auto_asc_desc")
+    optb3 = OptimalBinning(monotonic_trend="descending", verbose=True)
+
+    for optb in [optb0, optb1, optb2, optb3]:
+        optb.fit(x, y)
+        assert optb.status == "OPTIMAL"
+        assert optb.splits == approx([11.42500019, 12.32999992, 13.09499979,
+                                      13.70499992, 15.04500008, 16.92500019],
+                                     rel=1e-6)
+
+
+def test_numerical_min_max_n_bins():
+    optb_mip = OptimalBinning(solver="mip", min_n_bins=2, max_n_bins=5)
+    optb_cp = OptimalBinning(solver="cp", min_n_bins=2, max_n_bins=5)
+
+    for optb in [optb_mip, optb_cp]:
+        optb.fit(x, y)
+        assert optb.status == "OPTIMAL"
+        assert 2 <= len(optb.splits + 1) <= 5
+
+
+def test_outlier():
+    with raises(ValueError):
+        optb = OptimalBinning(outlier_detector="new_outlier")
+        optb.fit(x, y)
+
+    with raises(TypeError):
+        optb = OptimalBinning(outlier_detector="range", outlier_params=[])
+        optb.fit(x, y)
+
+    optb = OptimalBinning(outlier_detector="zscore", verbose=True)
+    optb.fit(x, y)
+    assert optb.splits == approx([11.42500019, 12.32999992, 13.09499979,
+                                  13.70499992, 15.04500008, 16.92500019],
+                                 rel=1e-6)
+
+    optb_eti = OptimalBinning(outlier_detector="range",
+                              outlier_params={"interval_length": 0.9,
+                                              "method": "ETI"})
+
+    optb_hdi = OptimalBinning(outlier_detector="range",
+                              outlier_params={"interval_length": 0.9,
+                                              "method": "HDI"})
+
+    for optb in [optb_eti, optb_hdi]:
+        optb.fit(x, y)
         assert optb.splits == approx([11.42500019, 12.32999992, 13.09499979,
                                       13.70499992, 15.04500008, 16.92500019],
                                      rel=1e-6)
@@ -207,6 +338,51 @@ def test_numerical_default_fit_transform():
     x_transform = optb.fit_transform(x, y, metric="woe")
     assert x_transform[:5] == approx([5.28332344, 5.28332344, 5.28332344,
                                       -3.12517033, 5.28332344], rel=1e-6)
+
+
+def test_categorical_transform():
+    x = np.array([
+        'Working', 'State servant', 'Working', 'Working', 'Working',
+        'State servant', 'Commercial associate', 'State servant',
+        'Pensioner', 'Working', 'Working', 'Pensioner', 'Working',
+        'Working', 'Working', 'Working', 'Working', 'Working', 'Working',
+        'State servant', 'Working', 'Commercial associate', 'Working',
+        'Pensioner', 'Working', 'Working', 'Working', 'Working',
+        'State servant', 'Working', 'Commercial associate', 'Working',
+        'Working', 'Commercial associate', 'State servant', 'Working',
+        'Commercial associate', 'Working', 'Pensioner', 'Working',
+        'Commercial associate', 'Working', 'Working', 'Pensioner',
+        'Working', 'Working', 'Pensioner', 'Working', 'State servant',
+        'Working', 'State servant', 'Commercial associate', 'Working',
+        'Commercial associate', 'Pensioner', 'Working', 'Pensioner',
+        'Working', 'Working', 'Working', 'Commercial associate', 'Working',
+        'Pensioner', 'Working', 'Commercial associate',
+        'Commercial associate', 'State servant', 'Working',
+        'Commercial associate', 'Commercial associate',
+        'Commercial associate', 'Working', 'Working', 'Working',
+        'Commercial associate', 'Working', 'Commercial associate',
+        'Working', 'Working', 'Pensioner', 'Working', 'Pensioner',
+        'Working', 'Working', 'Pensioner', 'Working', 'State servant',
+        'Working', 'Working', 'Working', 'Working', 'Working',
+        'Commercial associate', 'Commercial associate',
+        'Commercial associate', 'Working', 'Commercial associate',
+        'Working', 'Working', 'Pensioner'], dtype=object)
+
+    y = np.array([
+        1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0])
+
+    optb = OptimalBinning(dtype="categorical", solver="mip", cat_cutoff=0.1,
+                          verbose=True)
+    optb.fit(x, y)
+    x_transform = optb.transform(["Pensioner", "Working",
+                                  "Commercial associate", "State servant"])
+
+    assert x_transform == approx([-0.26662866, 0.30873548, -0.55431074,
+                                  0.30873548], rel=1e-6)
 
 
 def test_information():