+++ doc/source/release_notes.rst
@@ -1,10 +1,34 @@
 Release Notes
 =============
 
+Version 0.4.0 (2020-03-22)
+--------------------------
+
+New features:
+
+   - New ``monotonic_trend`` auto modes options: "auto_heuristic" and "auto_asc_desc".
+   - New ``monotonic_trend`` options: "peak_heuristic" and "valley_heuristic". These options produce a remarkable speedup for large size instances.
+   - Minimum Description Length Principle (MDLP) discretization algorithm.
+
+
+Improvements:
+
+   - ``BinningProcess`` now supports ``pandas.DataFrame`` as input X.
+   - New unit test added.
+
+
+Version 0.3.1 (2020-03-17)
+--------------------------
+
+Bugfixes:
+
+   - Fix setup.py packages using find_packages.
+
+
 Version 0.3.0 (2020-03-13)
 --------------------------
 
-New additions:
+New features:
 
    - Class ``OptBinning`` introduces a new constraint to reduce dominating bins, using parameter ``gamma``.
    - Metrics HHI, HHI regularized and Cramer's V added to ``binning_table.analysis`` method. Updated quality score.
@@ -20,10 +44,14 @@ Tutorials:
 Version 0.2.0 (2020-02-02)
 --------------------------
 
-New additions:
+New features:
 
    - Binning process to support optimal binning of all variables in dataset.
    - Added ``print_output`` option to ``binning_table.analysis`` method.
+
+
+Improvements:
+
    - New unit tests added.
 
 Tutorials: